pyproject.toml
[project]
name = "indiaquant-mcp"
version = "1.0.0"
description = "Real-time Indian stock market AI assistant using MCP + free APIs"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "yfinance>=0.2.36",
    "pandas>=2.1.0",
    "pandas-ta>=0.3.14b1",
    "scipy>=1.11.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.0",
    "cachetools>=5.3.0",
    "python-dotenv>=1.0.0",
    "vaderSentiment>=3.3.2",
    "numpy>=1.24.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
